Fixed environmental scaler
